Políticas de acceso

+++ app/Http/Controllers/Api/RecipeController.php
@@ -34,8 +34,12 @@ class RecipeController extends Controller
     }
 
     public function update(UpdateRecipeRequest $request, Recipe $recipe)
+
     {
+        $this->authorize('update', $recipe);
+
         $recipe->update($request->all());
+
 
         if ($tags = json_decode($request->tags)) {
             $recipe->tags()->attach($tags);
@@ -46,6 +50,8 @@ class RecipeController extends Controller
 
     public function destroy(Recipe $recipe)
     {
+        $this->authorize('delete', $recipe);
+
         $recipe->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
